chore: testing is more robust for main logic
